Move independent code to new lib folder
