+++ src/config/backgrounds.ts
@@ -1,4 +1,4 @@
-import { makeConfig } from '../helpers';
+import { makeConfig } from 'lib/helpers';
 
 export const backgroundConfig = makeConfig([
   {
@@ -29,7 +29,7 @@ export const backgroundConfig = makeConfig([
       id: 'satellite',
       type: 'raster',
       tiles: [
-        'https://tiles.maps.eox.at/wmts/1.0.0/s2cloudless-2020_3857/default/GoogleMapsCompatible/{z}/{y}/{x}.png'
+        'https://tiles.maps.eox.at/wmts/1.0.0/s2cloudless-2020_3857/default/GoogleMapsCompatible/{z}/{y}/{x}.png',
       ],
       tileSize: 256,
     },
