+++ src/lib/controls/CustomSlider.tsx
