+++ src/lib/controls/checkbox-tree/CheckboxTree.tsx
@@ -1,13 +1,13 @@
 import React, { useMemo } from 'react';
 import _ from 'lodash';
-
 import { TreeView } from '@mui/lab';
 import { ExpandMore as ExpandMoreIcon, ChevronRight as ChevronRightIcon } from '@mui/icons-material';
+import { useImmer } from 'use-immer';
+
+import { useChangeEffect } from 'lib/hooks/use-change-effect';
 
 import { dfs, getDescendants, TreeNode } from './tree-node';
 import { CheckboxTreeItem } from './CheckboxTreeItem';
-import { useImmer } from 'use-immer';
-import { useChangeEffect } from '../../hooks/use-change-effect';
 
 export interface CheckboxTreeConfig<T> {
   roots: TreeNode<T>[];
