+++ src/lib/controls/checkbox-tree/CheckboxTreeItem.tsx
@@ -1,5 +1,6 @@
 import { Checkbox, FormControlLabel } from '@mui/material';
 import { TreeItem } from '@mui/lab';
+
 import { CheckboxTreeState } from './CheckboxTree';
 import { TreeNode } from './tree-node';
