+++ src/lib/controls/use-data-params.ts
@@ -1,6 +1,21 @@
-import _ from 'lodash';
 import { useCallback, useState } from 'react';
-import { DataParamConfig, ParamDependencies, ParamDomains } from '../config/data/types';
+
+type Params = object;
+type ParamDomain<PT> = PT[];
+
+export type ParamDomains<PT extends Params> = {
+  [K in keyof PT]: ParamDomain<PT[K]>;
+};
+
+export type ParamDependencies<PT extends Params> = {
+  [K in keyof PT]?: (params: PT) => ParamDomain<PT[K]>;
+};
+
+export interface DataParamConfig<P extends Params> {
+  paramDomains: ParamDomains<P>;
+  paramDefaults: P;
+  paramDependencies?: ParamDependencies<P>;
+}
 
 function getNewParamOptions(updatedParams, domain, dependencyFn) {
   return dependencyFn?.(updatedParams) ?? domain;
