+++ src/lib/deck-layers/utils.ts
@@ -1,6 +1,6 @@
 import * as d3 from 'd3-scale';
-import { colorCssToRgb } from '../../helpers';
-import { VECTOR_COLOR_MAPS } from '../color-maps';
+import { colorCssToRgb } from 'lib/helpers';
+import { VECTOR_COLOR_MAPS } from '../../config/color-maps';
 
 export const mergeUpdateTriggers = (...propsArray) => {
   const updateTriggers = {};
