+++ src/lib/map/place-search/search-state.ts
@@ -1,5 +1,7 @@
 import { atom } from 'recoil';
 
+import { PlaceSearchResult } from 'lib/map/place-search/use-place-search';
+
 export const placeSearchActiveState = atom<boolean>({
   key: 'placeSearchActive',
   default: false,
@@ -9,20 +11,6 @@ export const placeSearchQueryState = atom<string>({
   key: 'placeSearchQuery',
   default: '',
 });
-
-export interface BoundingBox {
-  minX: number;
-  minY: number;
-  maxX: number;
-  maxY: number;
-}
-
-export interface PlaceSearchResult {
-  label: string;
-  latitude: number;
-  longitude: number;
-  boundingBox: BoundingBox;
-}
 
 export const placeSearchSelectedResultState = atom<PlaceSearchResult>({
   key: 'placeSearchSelectedResult',
