+++ tsconfig.json
@@ -21,7 +21,8 @@
     "noEmit": true,
     "jsx": "react-jsx",
     "noImplicitAny": false,
-    "downlevelIteration": true
+    "downlevelIteration": true,
+    "baseUrl": "./src"
   },
   "include": [
     "src"
